Fix Saturation Values
Anything above 1f would give over a full bar of saturation, which does not feel very intentional with the values here. If you use AppleSkin to view the saturation values ingame, most of the food items in this mod would give more saturation than can even fit on screen, which is obviously a problem.